+++ src/service/offer/reservation.ts
@@ -31,7 +31,7 @@ export type IAuthorizeSeatReservationResponse<T extends WebAPIIdentifier> =
 /**
  * 座席仮予約取消
  */
-export function voidTransaction(params: factory.task.IData<factory.taskName.VoidReserve>) {
+export function voidTransaction(params: factory.task.IData<factory.taskName.VoidReserveTransaction>) {
     return async (repos: {
         action: ActionRepo;
     }) => {