+++ src/service/task/confirmCancelReserve.ts
@@ -1,19 +1,18 @@
 import { IConnectionSettings, IOperation } from '../task';
 
 import * as factory from '../../factory';
-
 import { MongoRepository as ActionRepo } from '../../repo/action';
 
-import * as ReservationOfferService from '../offer/reservation';
+import * as ReservationService from '../reservation';
 
 /**
  * タスク実行関数
  */
-export function call(data: factory.task.IData<factory.taskName.VoidReserve>): IOperation<void> {
+export function call(data: factory.task.IData<factory.taskName.ConfirmCancelReserve>): IOperation<void> {
     return async (settings: IConnectionSettings) => {
         const actionRepo = new ActionRepo(settings.connection);
 
-        await ReservationOfferService.voidTransaction(data)({
+        await ReservationService.cancelReservation(data)({
             action: actionRepo
         });
     };