+++ src/service/task/confirmPay.ts
@@ -1,18 +1,19 @@
 import { IConnectionSettings, IOperation } from '../task';
 
 import * as factory from '../../factory';
+
 import { MongoRepository as ActionRepo } from '../../repo/action';
 
-import * as ReservationService from '../reservation';
+import * as PaymentService from '../payment';
 
 /**
  * タスク実行関数
  */
-export function call(data: factory.task.IData<factory.taskName.CancelReservation>): IOperation<void> {
+export function call(data: factory.task.IData<factory.taskName.ConfirmPay>): IOperation<void> {
     return async (settings: IConnectionSettings) => {
         const actionRepo = new ActionRepo(settings.connection);
 
-        await ReservationService.cancelReservation(data)({
+        await PaymentService.pay(data)({
             action: actionRepo
         });
     };