+++ src/service/task/voidMoneyTransferTransaction.ts
@@ -5,17 +5,17 @@ import * as factory from '../../factory';
 import { MongoRepository as ActionRepo } from '../../repo/action';
 import { MongoRepository as TransactionRepo } from '../../repo/transaction';
 
-import * as PaymentService from '../payment';
+import * as OfferService from '../offer';
 
 /**
  * タスク実行関数
  */
-export function call(data: factory.task.IData<factory.taskName.VoidPayment>): IOperation<void> {
+export function call(data: factory.task.IData<factory.taskName.VoidMoneyTransferTransaction>): IOperation<void> {
     return async (settings: IConnectionSettings) => {
         const actionRepo = new ActionRepo(settings.connection);
         const transactionRepo = new TransactionRepo(settings.connection);
 
-        await PaymentService.voidPayment(data)({
+        await OfferService.monetaryAmount.voidTransaction(data)({
             action: actionRepo,
             transaction: transactionRepo
         });